🔥 remove: 테스트 페이지 삭제 (Jira Ticket : CCC-83)
